requirements.txt
# ───────────── FastAPI Framework ─────────────
fastapi==0.115.14
uvicorn[standard]==0.29.0

# ───────────── Machine Learning ─────────────
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
xgboost==2.0.3
lightgbm==4.3.0
scikit-learn==1.3.2
imbalanced-learn==0.11.0
# scikit-learn==1.6.1
# imbalanced-learn==0.12.0


# ───────────── Auth & Security ─────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# ───────────── Config & Env Management ─────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ───────────── HTTP Clients ────────────────
httpx==0.27.0

# ───────────── MongoDB Driver ──────────────
motor>=3.0

# ───────────── Notes ───────────────────────
# To install: pip install -r requirements.txt
